Login Funktion

+++ src/routes/+page.server.js
@@ -1,6 +1,7 @@
 import { fail } from '@sveltejs/kit';
 import { ObjectId } from 'mongodb';
 import { getCardsCollection } from '$lib/server/mongo';
+import { requireUser } from '$lib/server/auth';
 
 function mapCard(doc) {
 	return {
@@ -16,16 +17,18 @@ function mapCard(doc) {
 	};
 }
 
-export async function load() {
+export async function load({ locals }) {
+	const user = requireUser(locals);
 	const collection = await getCardsCollection();
-	const docs = await collection.find({}).sort({ createdAt: -1 }).toArray();
+	const docs = await collection.find({ userId: user }).sort({ createdAt: -1 }).toArray();
 	return {
 		cards: docs.map(mapCard)
 	};
 }
 
 export const actions = {
-	delete: async ({ request }) => {
+	delete: async ({ request, locals }) => {
+		const user = requireUser(locals);
 		const data = await request.formData();
 		const id = data.get('id');
 
@@ -37,7 +40,7 @@ export const actions = {
 		}
 
 		const collection = await getCardsCollection();
-		await collection.deleteOne({ _id: new ObjectId(id) });
+		await collection.deleteOne({ _id: new ObjectId(id), userId: user });
 		return { success: true };
 	}
 };
